+++ backend/model/video_comment.go
@@ -7,11 +7,11 @@ import (
 
 type VideoComment struct {
 	// Id, VideoId, UserId, Content, QuoteId, CommentTime, DeleteTime
-	Id          int64        `default:"0"`
-	VideoId     int          `default:"0"`
-	UserId      int          `default:"0"`
+	Id          uint         `default:"0"`
+	VideoId     uint         `default:"0"`
+	UserId      uint         `default:"0"`
 	Content     string       `default:""`
-	QuoteId     int64        `default:"0"`
+	QuoteId     uint         `default:"0"`
 	CommentTime time.Time    `default:"0000-00-00 00:00:00"`
 	DeleteTime  sql.NullTime `default:"0000-00-00 00:00:00"`
 }
